com.github.jackalsin.leetcode.problems/src/test/java/_0801_0850/_820_Short_Encoding_of_Words/SolutionITest.java
package _0801_0850._820_Short_Encoding_of_Words;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 9/1/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final String[] input = {"time", "me", "bell"};
    assertEquals(10, solution.minimumLengthEncoding(input));
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}